Make Exposure Summary Table

Generates the summary exposure table with final factor scores, standard deviations, and distribution of spatial and temporal for each species, as well as final exposure scores. One table is generated per species and saved in the species folder. Requires directory to be set up per directions in the package documentation/manual.

Usage

make_exposure_table(
  species,
  present_time,
  future_time,
  variable_df,
  model_metrics,
  auc_col,
  model_confidence
)

Arguments

species

names of the species to plot. Must match folder name to pull correct data and save table correctly.

present_time, future_time

character strings indicating the present and future time series to compare. Example: '1993-2019'. Used to pull correct calculations and save the data properly

variable_df

a data.frame containing all possible environmental variables, such as from the MOM6 model. Must contain columns Long.Name and Short.Name, containing the full names and abbreviated names of the variables. Abbreviated names should correspond to those in the weights vector produced by normalize_variable_weights

model_metrics

a data.table containing the model evaluation metrics for each species. Generated by make_evaluation_csv.

auc_col

name of the column containing the AUC values to be included in the table footnote

model_confidence

a data.table containing the model confidence scores for each species.

Value

Function does not return anything. The table is saved to species- and timeframe-specific figures folder. See manual for directory set up.