Make Sensitivity Table

Generates the sensitivity table with final attribute scores, data quality, and distribution of tallies for each species. One table is generated per species and saved in the designated directory. Requires directory to be set up per directions in the package documentation/manual.

Usage

make_sensitivity_table(
  species,
  species_col,
  attribute_names_raw,
  attribute_names_clean,
  total_sens_col,
  certainty_col,
  raw_data,
  sensitivity,
  data_quality,
  table_dir
)

Arguments

species

names of species to produce table for. Must match Stock.Name in raw_data

species_col

name of column with species names to help subset data. Should be the same across raw_data, sensitivity, and data_quality

attribute_names_raw, attribute_names_clean

vectors of raw attribute names to help subset sensitivity and data_quality, and cleaned up names to use in table

total_sens_col, certainty_col

names of column containing total sensitivity scores and associated certainty in sensitivity

raw_data

raw data exported from FCVA database, including the column Stock.Name to subset data to correct species

sensitivity

TRUE indicates sensitivity attribute scores should be plotted. FALSE indicates exposure factor scores should be plotted

data_quality

data.frame with data quality scores for each attribute and species. Generated by data_quality.score. Must include the column Stock.Name to subset data to correct species

table_dir

file path to folder to save tables in

Value

No returns. Saves pdf file containing the sensitivity table containing the expert scores, data quality, and overall sensitivity score, with certainty.